refactor names entities

+++ src/contracts/contract_shared/responsible.v
@@ -1,6 +1,6 @@
 module contract_shared
 
-pub enum Responsible as int {
+pub enum Responsible as i8 {
 	pai = 0
 	mae = 1
 }
